backend/zchat/requirements.txt
asgiref
Django
django-cors-headers
django-allauth
djangorestframework
djangorestframework-simplejwt
PyJWT
environs
django-environ
psycopg2-binary
python-dotenv
django-channels
daphne
pillow